backend/requirements.txt
# path: backend/requirements.txt
fastapi==0.112.0
httpx==0.27.0
uvicorn==0.30.0
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic-settings==2.4.0
requests
psycopg2-binary
